Facebook iOS SDK 3.8
Summary:
     see https://developers.facebook.com/ios/change-log-3.x/

    Upgrading from 3.7
     https://developers.facebook.com/docs/ios/upgrading-from-3.7-to-3.8/

Reviewed By: chrisp

Test Plan: Release Testing.

+++ src/FBAppBridgeTypeToJSONConverter.m
@@ -122,7 +122,7 @@ static NSString *const FBAppBridgeTypeIdentifier = @"com.facebook.Facebook.FBApp
 }
 
 - (NSArray *)convertedArrayFromArray:(NSArray *)array convertingToJSON:(BOOL)convertingToJSON {
-    int length = array.count;
+    NSUInteger length = array.count;
     if (length == 0) {
         return array;
     }
