+++ src/FBRequestBody.m
@@ -88,7 +88,7 @@ static NSString *kStringBoundary = @"3i2ndDfv2rTHiSisAbouNdArYfORhtTPEefj3q2f";
     NSData *data = UIImageJPEGRepresentation(image, [FBSettings defaultJPEGCompressionQuality]);
     [self.mutableData appendData:data];
     [self appendRecordBoundary];
-    [logger appendFormat:@"\n    %@:\t<Image - %d kB>", key, [data length] / 1024];
+    [logger appendFormat:@"\n    %@:\t<Image - %lu kB>", key, (unsigned long)([data length] / 1024)];
 }
 
 - (void)appendWithKey:(NSString *)key
@@ -101,7 +101,7 @@ static NSString *kStringBoundary = @"3i2ndDfv2rTHiSisAbouNdArYfORhtTPEefj3q2f";
     [self appendUTF8:@"Content-Type: content/unknown\r\n\r\n"];
     [self.mutableData appendData:data];
     [self appendRecordBoundary];
-    [logger appendFormat:@"\n    %@:\t<Data - %d kB>", key, [data length] / 1024];
+    [logger appendFormat:@"\n    %@:\t<Data - %lu kB>", key, (unsigned long)([data length] / 1024)];
 }
 
 - (NSData *)data
